@@ -25,16 +25,16 @@
 
 
     "require": {
-        "php": "~7.0"
+        "php": "~7.1"
     },
 
     "require-dev": {
-        "codeception/codeception": "~2.2",
-        "phpunit/php-code-coverage": "~3.0",
+        "codeception/codeception": "~2.5",
+        "phpunit/php-code-coverage": "~6.0",
 
-        "squizlabs/php_codesniffer": "~2.5",
+        "squizlabs/php_codesniffer": "~3.4",
 
-        "twig/twig": "~1.24"
+        "twig/twig": "~2.0"
     },
 
     "suggest": {
