@@ -1,12 +1,9 @@
-<?php  //[STAMP] 16758fb987cbb3c846d5a082e48c1065
+<?php  //[STAMP] 6b282b3f114eb841bcfbd62136f0fe05
 namespace _generated;
 
 // This class was automatically generated by build task
 // You should not change it manually as it will be overwritten on next build
 // @codingStandardsIgnoreFile
-
-use Codeception\Module\PhpBrowser;
-use Helper\Acceptance;
 
 trait AcceptanceTesterActions
 {
@@ -19,49 +16,14 @@ trait AcceptanceTesterActions
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * Sets the HTTP header to the passed value - which is used on
-     * subsequent HTTP requests through PhpBrowser.
+     * Alias to `haveHttpHeader`
      *
-     * Example:
-     * ```php
-     * <?php
-     * $I->setHeader('X-Requested-With', 'Codeception');
-     * $I->amOnPage('test-headers.php');
-     * ?>
-     * ```
-     *
-     * @param string $name the name of the request header
-     * @param string $value the value to set it to for subsequent
-     *        requests
+     * @param $name
+     * @param $value
      * @see \Codeception\Module\PhpBrowser::setHeader()
      */
     public function setHeader($name, $value) {
         return $this->getScenario()->runStep(new \Codeception\Step\Action('setHeader', func_get_args()));
-    }
-
- 
-    /**
-     * [!] Method is generated. Documentation taken from corresponding module.
-     *
-     * Deletes the header with the passed name.  Subsequent requests
-     * will not have the deleted header in its request.
-     *
-     * Example:
-     * ```php
-     * <?php
-     * $I->setHeader('X-Requested-With', 'Codeception');
-     * $I->amOnPage('test-headers.php');
-     * // ...
-     * $I->deleteHeader('X-Requested-With');
-     * $I->amOnPage('some-other-page.php');
-     * ?>
-     * ```
-     * 
-     * @param string $name the name of the header to delete.
-     * @see \Codeception\Module\PhpBrowser::deleteHeader()
-     */
-    public function deleteHeader($name) {
-        return $this->getScenario()->runStep(new \Codeception\Step\Action('deleteHeader', func_get_args()));
     }
 
  
@@ -155,6 +117,66 @@ trait AcceptanceTesterActions
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
+     * Sets the HTTP header to the passed value - which is used on
+     * subsequent HTTP requests through PhpBrowser.
+     *
+     * Example:
+     * ```php
+     * <?php
+     * $I->haveHttpHeader('X-Requested-With', 'Codeception');
+     * $I->amOnPage('test-headers.php');
+     * ?>
+     * ```
+     *
+     * To use special chars in Header Key use HTML Character Entities:
+     * Example:
+     * Header with underscore - 'Client_Id'
+     * should be represented as - 'Client&#x0005F;Id' or 'Client&#95;Id'
+     *
+     * ```php
+     * <?php
+     * $I->haveHttpHeader('Client&#95;Id', 'Codeception');
+     * ?>
+     * ```
+     *
+     * @param string $name the name of the request header
+     * @param string $value the value to set it to for subsequent
+     *        requests
+     * @see \Codeception\Lib\InnerBrowser::haveHttpHeader()
+     */
+    public function haveHttpHeader($name, $value) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('haveHttpHeader', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Deletes the header with the passed name.  Subsequent requests
+     * will not have the deleted header in its request.
+     *
+     * Example:
+     * ```php
+     * <?php
+     * $I->haveHttpHeader('X-Requested-With', 'Codeception');
+     * $I->amOnPage('test-headers.php');
+     * // ...
+     * $I->deleteHeader('X-Requested-With');
+     * $I->amOnPage('some-other-page.php');
+     * ?>
+     * ```
+     *
+     * @param string $name the name of the header to delete.
+     * @see \Codeception\Lib\InnerBrowser::deleteHeader()
+     */
+    public function deleteHeader($name) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('deleteHeader', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
      * Opens the page for the given relative URI.
      *
      * ``` php
@@ -163,10 +185,9 @@ trait AcceptanceTesterActions
      * $I->amOnPage('/');
      * // opens /register page
      * $I->amOnPage('/register');
-     * ?>
      * ```
      *
-     * @param $page
+     * @param string $page
      * @see \Codeception\Lib\InnerBrowser::amOnPage()
      */
     public function amOnPage($page) {
@@ -196,7 +217,7 @@ trait AcceptanceTesterActions
      * // CSS button
      * $I->click('#form input[type=submit]');
      * // XPath
-     * $I->click('//form/*[@type=submit]');
+     * $I->click('//form/*[@type="submit"]');
      * // link in context
      * $I->click('Logout', '#nav');
      * // using strict locator
@@ -216,19 +237,35 @@ trait AcceptanceTesterActions
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * Checks that the current page contains the given string.
-     * Specify a locator as the second parameter to match a specific region.
+     * Checks that the current page contains the given string (case insensitive).
+     *
+     * You can specify a specific HTML element (via CSS or XPath) as the second
+     * parameter to only search within that element.
      *
      * ``` php
      * <?php
-     * $I->see('Logout'); // I can suppose user is logged in
-     * $I->see('Sign Up','h1'); // I can suppose it's a signup page
-     * $I->see('Sign Up','//body/h1'); // with XPath
-     * ?>
+     * $I->see('Logout');                        // I can suppose user is logged in
+     * $I->see('Sign Up', 'h1');                 // I can suppose it's a signup page
+     * $I->see('Sign Up', '//body/h1');          // with XPath
+     * $I->see('Sign Up', ['css' => 'body h1']); // with strict CSS locator
      * ```
      *
-     * @param      $text
-     * @param null $selector
+     * Note that the search is done after stripping all HTML tags from the body,
+     * so `$I->see('strong')` will return true for strings like:
+     *
+     *   - `<p>I am Stronger than thou</p>`
+     *   - `<script>document.createElement('strong');</script>`
+     *
+     * But will *not* be true for strings like:
+     *
+     *   - `<strong>Home</strong>`
+     *   - `<div class="strong">Home</strong>`
+     *   - `<!-- strong -->`
+     *
+     * For checking the raw source code, use `seeInSource()`.
+     *
+     * @param string $text
+     * @param array|string $selector optional
      * Conditional Assertion: Test won't be stopped on fail
      * @see \Codeception\Lib\InnerBrowser::see()
      */
@@ -238,19 +275,35 @@ trait AcceptanceTesterActions
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * Checks that the current page contains the given string.
-     * Specify a locator as the second parameter to match a specific region.
+     * Checks that the current page contains the given string (case insensitive).
+     *
+     * You can specify a specific HTML element (via CSS or XPath) as the second
+     * parameter to only search within that element.
      *
      * ``` php
      * <?php
-     * $I->see('Logout'); // I can suppose user is logged in
-     * $I->see('Sign Up','h1'); // I can suppose it's a signup page
-     * $I->see('Sign Up','//body/h1'); // with XPath
-     * ?>
+     * $I->see('Logout');                        // I can suppose user is logged in
+     * $I->see('Sign Up', 'h1');                 // I can suppose it's a signup page
+     * $I->see('Sign Up', '//body/h1');          // with XPath
+     * $I->see('Sign Up', ['css' => 'body h1']); // with strict CSS locator
      * ```
      *
-     * @param      $text
-     * @param null $selector
+     * Note that the search is done after stripping all HTML tags from the body,
+     * so `$I->see('strong')` will return true for strings like:
+     *
+     *   - `<p>I am Stronger than thou</p>`
+     *   - `<script>document.createElement('strong');</script>`
+     *
+     * But will *not* be true for strings like:
+     *
+     *   - `<strong>Home</strong>`
+     *   - `<div class="strong">Home</strong>`
+     *   - `<!-- strong -->`
+     *
+     * For checking the raw source code, use `seeInSource()`.
+     *
+     * @param string $text
+     * @param array|string $selector optional
      * @see \Codeception\Lib\InnerBrowser::see()
      */
     public function see($text, $selector = null) {
@@ -261,19 +314,33 @@ trait AcceptanceTesterActions
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * Checks that the current page doesn't contain the text specified.
+     * Checks that the current page doesn't contain the text specified (case insensitive).
      * Give a locator as the second parameter to match a specific region.
      *
      * ```php
      * <?php
-     * $I->dontSee('Login'); // I can suppose user is already logged in
-     * $I->dontSee('Sign Up','h1'); // I can suppose it's not a signup page
-     * $I->dontSee('Sign Up','//body/h1'); // with XPath
-     * ?>
+     * $I->dontSee('Login');                         // I can suppose user is already logged in
+     * $I->dontSee('Sign Up','h1');                  // I can suppose it's not a signup page
+     * $I->dontSee('Sign Up','//body/h1');           // with XPath
+     * $I->dontSee('Sign Up', ['css' => 'body h1']); // with strict CSS locator
      * ```
      *
-     * @param      $text
-     * @param null $selector
+     * Note that the search is done after stripping all HTML tags from the body,
+     * so `$I->dontSee('strong')` will fail on strings like:
+     *
+     *   - `<p>I am Stronger than thou</p>`
+     *   - `<script>document.createElement('strong');</script>`
+     *
+     * But will ignore strings like:
+     *
+     *   - `<strong>Home</strong>`
+     *   - `<div class="strong">Home</strong>`
+     *   - `<!-- strong -->`
+     *
+     * For checking the raw source code, use `seeInSource()`.
+     *
+     * @param string $text
+     * @param array|string $selector optional
      * Conditional Assertion: Test won't be stopped on fail
      * @see \Codeception\Lib\InnerBrowser::dontSee()
      */
@@ -283,23 +350,111 @@ trait AcceptanceTesterActions
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * Checks that the current page doesn't contain the text specified.
+     * Checks that the current page doesn't contain the text specified (case insensitive).
      * Give a locator as the second parameter to match a specific region.
      *
      * ```php
      * <?php
-     * $I->dontSee('Login'); // I can suppose user is already logged in
-     * $I->dontSee('Sign Up','h1'); // I can suppose it's not a signup page
-     * $I->dontSee('Sign Up','//body/h1'); // with XPath
-     * ?>
+     * $I->dontSee('Login');                         // I can suppose user is already logged in
+     * $I->dontSee('Sign Up','h1');                  // I can suppose it's not a signup page
+     * $I->dontSee('Sign Up','//body/h1');           // with XPath
+     * $I->dontSee('Sign Up', ['css' => 'body h1']); // with strict CSS locator
      * ```
      *
-     * @param      $text
-     * @param null $selector
+     * Note that the search is done after stripping all HTML tags from the body,
+     * so `$I->dontSee('strong')` will fail on strings like:
+     *
+     *   - `<p>I am Stronger than thou</p>`
+     *   - `<script>document.createElement('strong');</script>`
+     *
+     * But will ignore strings like:
+     *
+     *   - `<strong>Home</strong>`
+     *   - `<div class="strong">Home</strong>`
+     *   - `<!-- strong -->`
+     *
+     * For checking the raw source code, use `seeInSource()`.
+     *
+     * @param string $text
+     * @param array|string $selector optional
      * @see \Codeception\Lib\InnerBrowser::dontSee()
      */
     public function dontSee($text, $selector = null) {
         return $this->getScenario()->runStep(new \Codeception\Step\Assertion('dontSee', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that the current page contains the given string in its
+     * raw source code.
+     *
+     * ``` php
+     * <?php
+     * $I->seeInSource('<h1>Green eggs &amp; ham</h1>');
+     * ```
+     *
+     * @param      $raw
+     * Conditional Assertion: Test won't be stopped on fail
+     * @see \Codeception\Lib\InnerBrowser::seeInSource()
+     */
+    public function canSeeInSource($raw) {
+        return $this->getScenario()->runStep(new \Codeception\Step\ConditionalAssertion('seeInSource', func_get_args()));
+    }
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that the current page contains the given string in its
+     * raw source code.
+     *
+     * ``` php
+     * <?php
+     * $I->seeInSource('<h1>Green eggs &amp; ham</h1>');
+     * ```
+     *
+     * @param      $raw
+     * @see \Codeception\Lib\InnerBrowser::seeInSource()
+     */
+    public function seeInSource($raw) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Assertion('seeInSource', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that the current page contains the given string in its
+     * raw source code.
+     *
+     * ```php
+     * <?php
+     * $I->dontSeeInSource('<h1>Green eggs &amp; ham</h1>');
+     * ```
+     *
+     * @param      $raw
+     * Conditional Assertion: Test won't be stopped on fail
+     * @see \Codeception\Lib\InnerBrowser::dontSeeInSource()
+     */
+    public function cantSeeInSource($raw) {
+        return $this->getScenario()->runStep(new \Codeception\Step\ConditionalAssertion('dontSeeInSource', func_get_args()));
+    }
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that the current page contains the given string in its
+     * raw source code.
+     *
+     * ```php
+     * <?php
+     * $I->dontSeeInSource('<h1>Green eggs &amp; ham</h1>');
+     * ```
+     *
+     * @param      $raw
+     * @see \Codeception\Lib\InnerBrowser::dontSeeInSource()
+     */
+    public function dontSeeInSource($raw) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Assertion('dontSeeInSource', func_get_args()));
     }
 
  
@@ -316,8 +471,8 @@ trait AcceptanceTesterActions
      * ?>
      * ```
      *
-     * @param      $text
-     * @param null $url
+     * @param string $text
+     * @param string $url optional
      * Conditional Assertion: Test won't be stopped on fail
      * @see \Codeception\Lib\InnerBrowser::seeLink()
      */
@@ -337,8 +492,8 @@ trait AcceptanceTesterActions
      * ?>
      * ```
      *
-     * @param      $text
-     * @param null $url
+     * @param string $text
+     * @param string $url optional
      * @see \Codeception\Lib\InnerBrowser::seeLink()
      */
     public function seeLink($text, $url = null) {
@@ -359,8 +514,8 @@ trait AcceptanceTesterActions
      * ?>
      * ```
      *
-     * @param $text
-     * @param null $url
+     * @param string $text
+     * @param string $url optional
      * Conditional Assertion: Test won't be stopped on fail
      * @see \Codeception\Lib\InnerBrowser::dontSeeLink()
      */
@@ -380,8 +535,8 @@ trait AcceptanceTesterActions
      * ?>
      * ```
      *
-     * @param $text
-     * @param null $url
+     * @param string $text
+     * @param string $url optional
      * @see \Codeception\Lib\InnerBrowser::dontSeeLink()
      */
     public function dontSeeLink($text, $url = null) {
@@ -403,7 +558,7 @@ trait AcceptanceTesterActions
      * ?>
      * ```
      *
-     * @param $uri
+     * @param string $uri
      * Conditional Assertion: Test won't be stopped on fail
      * @see \Codeception\Lib\InnerBrowser::seeInCurrentUrl()
      */
@@ -424,7 +579,7 @@ trait AcceptanceTesterActions
      * ?>
      * ```
      *
-     * @param $uri
+     * @param string $uri
      * @see \Codeception\Lib\InnerBrowser::seeInCurrentUrl()
      */
     public function seeInCurrentUrl($uri) {
@@ -443,7 +598,7 @@ trait AcceptanceTesterActions
      * ?>
      * ```
      *
-     * @param $uri
+     * @param string $uri
      * Conditional Assertion: Test won't be stopped on fail
      * @see \Codeception\Lib\InnerBrowser::dontSeeInCurrentUrl()
      */
@@ -461,7 +616,7 @@ trait AcceptanceTesterActions
      * ?>
      * ```
      *
-     * @param $uri
+     * @param string $uri
      * @see \Codeception\Lib\InnerBrowser::dontSeeInCurrentUrl()
      */
     public function dontSeeInCurrentUrl($uri) {
@@ -482,7 +637,7 @@ trait AcceptanceTesterActions
      * ?>
      * ```
      *
-     * @param $uri
+     * @param string $uri
      * Conditional Assertion: Test won't be stopped on fail
      * @see \Codeception\Lib\InnerBrowser::seeCurrentUrlEquals()
      */
@@ -502,7 +657,7 @@ trait AcceptanceTesterActions
      * ?>
      * ```
      *
-     * @param $uri
+     * @param string $uri
      * @see \Codeception\Lib\InnerBrowser::seeCurrentUrlEquals()
      */
     public function seeCurrentUrlEquals($uri) {
@@ -523,7 +678,7 @@ trait AcceptanceTesterActions
      * ?>
      * ```
      *
-     * @param $uri
+     * @param string $uri
      * Conditional Assertion: Test won't be stopped on fail
      * @see \Codeception\Lib\InnerBrowser::dontSeeCurrentUrlEquals()
      */
@@ -543,7 +698,7 @@ trait AcceptanceTesterActions
      * ?>
      * ```
      *
-     * @param $uri
+     * @param string $uri
      * @see \Codeception\Lib\InnerBrowser::dontSeeCurrentUrlEquals()
      */
     public function dontSeeCurrentUrlEquals($uri) {
@@ -559,11 +714,11 @@ trait AcceptanceTesterActions
      * ``` php
      * <?php
      * // to match root url
-     * $I->seeCurrentUrlMatches('~$/users/(\d+)~');
+     * $I->seeCurrentUrlMatches('~^/users/(\d+)~');
      * ?>
      * ```
      *
-     * @param $uri
+     * @param string $uri
      * Conditional Assertion: Test won't be stopped on fail
      * @see \Codeception\Lib\InnerBrowser::seeCurrentUrlMatches()
      */
@@ -578,11 +733,11 @@ trait AcceptanceTesterActions
      * ``` php
      * <?php
      * // to match root url
-     * $I->seeCurrentUrlMatches('~$/users/(\d+)~');
+     * $I->seeCurrentUrlMatches('~^/users/(\d+)~');
      * ?>
      * ```
      *
-     * @param $uri
+     * @param string $uri
      * @see \Codeception\Lib\InnerBrowser::seeCurrentUrlMatches()
      */
     public function seeCurrentUrlMatches($uri) {
@@ -598,11 +753,11 @@ trait AcceptanceTesterActions
      * ``` php
      * <?php
      * // to match root url
-     * $I->dontSeeCurrentUrlMatches('~$/users/(\d+)~');
+     * $I->dontSeeCurrentUrlMatches('~^/users/(\d+)~');
      * ?>
      * ```
      *
-     * @param $uri
+     * @param string $uri
      * Conditional Assertion: Test won't be stopped on fail
      * @see \Codeception\Lib\InnerBrowser::dontSeeCurrentUrlMatches()
      */
@@ -617,11 +772,11 @@ trait AcceptanceTesterActions
      * ``` php
      * <?php
      * // to match root url
-     * $I->dontSeeCurrentUrlMatches('~$/users/(\d+)~');
+     * $I->dontSeeCurrentUrlMatches('~^/users/(\d+)~');
      * ?>
      * ```
      *
-     * @param $uri
+     * @param string $uri
      * @see \Codeception\Lib\InnerBrowser::dontSeeCurrentUrlMatches()
      */
     public function dontSeeCurrentUrlMatches($uri) {
@@ -632,19 +787,18 @@ trait AcceptanceTesterActions
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * Executes the given regular expression against the current URI and returns the first match.
+     * Executes the given regular expression against the current URI and returns the first capturing group.
      * If no parameters are provided, the full URI is returned.
      *
      * ``` php
      * <?php
-     * $user_id = $I->grabFromCurrentUrl('~$/user/(\d+)/~');
+     * $user_id = $I->grabFromCurrentUrl('~^/user/(\d+)/~');
      * $uri = $I->grabFromCurrentUrl();
      * ?>
      * ```
      *
-     * @param null $uri
+     * @param string $uri optional
      *
-     * @internal param $url
      * @return mixed
      * @see \Codeception\Lib\InnerBrowser::grabFromCurrentUrl()
      */
@@ -736,8 +890,8 @@ trait AcceptanceTesterActions
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * Checks that the given input field or textarea contains the given value.
-     * For fuzzy locators, fields are matched by label text, the "name" attribute, CSS, and XPath.
+     * Checks that the given input field or textarea *equals* (i.e. not just contains) the given value.
+     * Fields are matched by label text, the "name" attribute, CSS, or XPath.
      *
      * ``` php
      * <?php
@@ -761,8 +915,8 @@ trait AcceptanceTesterActions
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * Checks that the given input field or textarea contains the given value.
-     * For fuzzy locators, fields are matched by label text, the "name" attribute, CSS, and XPath.
+     * Checks that the given input field or textarea *equals* (i.e. not just contains) the given value.
+     * Fields are matched by label text, the "name" attribute, CSS, or XPath.
      *
      * ``` php
      * <?php
@@ -840,7 +994,7 @@ trait AcceptanceTesterActions
      *
      * Checks if the array of form parameters (name => value) are set on the form matched with the
      * passed selector.
-     * 
+     *
      * ``` php
      * <?php
      * $I->seeInFormFields('form[name=myform]', [
@@ -849,10 +1003,10 @@ trait AcceptanceTesterActions
      * ]);
      * ?>
      * ```
-     * 
+     *
      * For multi-select elements, or to check values of multiple elements with the same name, an
      * array may be passed:
-     * 
+     *
      * ``` php
      * <?php
      * $I->seeInFormFields('.form-class', [
@@ -869,7 +1023,7 @@ trait AcceptanceTesterActions
      * ```
      *
      * Additionally, checkbox values can be checked with a boolean.
-     * 
+     *
      * ``` php
      * <?php
      * $I->seeInFormFields('#form-id', [
@@ -878,9 +1032,9 @@ trait AcceptanceTesterActions
      * ]);
      * ?>
      * ```
-     * 
+     *
      * Pair this with submitForm for quick testing magic.
-     * 
+     *
      * ``` php
      * <?php
      * $form = [
@@ -894,7 +1048,7 @@ trait AcceptanceTesterActions
      * $I->seeInFormFields('//form[@id=my-form]', $form);
      * ?>
      * ```
-     * 
+     *
      * @param $formSelector
      * @param $params
      * Conditional Assertion: Test won't be stopped on fail
@@ -908,7 +1062,7 @@ trait AcceptanceTesterActions
      *
      * Checks if the array of form parameters (name => value) are set on the form matched with the
      * passed selector.
-     * 
+     *
      * ``` php
      * <?php
      * $I->seeInFormFields('form[name=myform]', [
@@ -917,10 +1071,10 @@ trait AcceptanceTesterActions
      * ]);
      * ?>
      * ```
-     * 
+     *
      * For multi-select elements, or to check values of multiple elements with the same name, an
      * array may be passed:
-     * 
+     *
      * ``` php
      * <?php
      * $I->seeInFormFields('.form-class', [
@@ -937,7 +1091,7 @@ trait AcceptanceTesterActions
      * ```
      *
      * Additionally, checkbox values can be checked with a boolean.
-     * 
+     *
      * ``` php
      * <?php
      * $I->seeInFormFields('#form-id', [
@@ -946,9 +1100,9 @@ trait AcceptanceTesterActions
      * ]);
      * ?>
      * ```
-     * 
+     *
      * Pair this with submitForm for quick testing magic.
-     * 
+     *
      * ``` php
      * <?php
      * $form = [
@@ -962,7 +1116,7 @@ trait AcceptanceTesterActions
      * $I->seeInFormFields('//form[@id=my-form]', $form);
      * ?>
      * ```
-     * 
+     *
      * @param $formSelector
      * @param $params
      * @see \Codeception\Lib\InnerBrowser::seeInFormFields()
@@ -977,7 +1131,7 @@ trait AcceptanceTesterActions
      *
      * Checks if the array of form parameters (name => value) are not set on the form matched with
      * the passed selector.
-     * 
+     *
      * ``` php
      * <?php
      * $I->dontSeeInFormFields('form[name=myform]', [
@@ -986,10 +1140,10 @@ trait AcceptanceTesterActions
      * ]);
      * ?>
      * ```
-     * 
+     *
      * To check that an element hasn't been assigned any one of many values, an array can be passed
      * as the value:
-     * 
+     *
      * ``` php
      * <?php
      * $I->dontSeeInFormFields('.form-class', [
@@ -1002,7 +1156,7 @@ trait AcceptanceTesterActions
      * ```
      *
      * Additionally, checkbox values can be checked with a boolean.
-     * 
+     *
      * ``` php
      * <?php
      * $I->dontSeeInFormFields('#form-id', [
@@ -1011,7 +1165,7 @@ trait AcceptanceTesterActions
      * ]);
      * ?>
      * ```
-     * 
+     *
      * @param $formSelector
      * @param $params
      * Conditional Assertion: Test won't be stopped on fail
@@ -1025,7 +1179,7 @@ trait AcceptanceTesterActions
      *
      * Checks if the array of form parameters (name => value) are not set on the form matched with
      * the passed selector.
-     * 
+     *
      * ``` php
      * <?php
      * $I->dontSeeInFormFields('form[name=myform]', [
@@ -1034,10 +1188,10 @@ trait AcceptanceTesterActions
      * ]);
      * ?>
      * ```
-     * 
+     *
      * To check that an element hasn't been assigned any one of many values, an array can be passed
      * as the value:
-     * 
+     *
      * ``` php
      * <?php
      * $I->dontSeeInFormFields('.form-class', [
@@ -1050,7 +1204,7 @@ trait AcceptanceTesterActions
      * ```
      *
      * Additionally, checkbox values can be checked with a boolean.
-     * 
+     *
      * ``` php
      * <?php
      * $I->dontSeeInFormFields('#form-id', [
@@ -1059,7 +1213,7 @@ trait AcceptanceTesterActions
      * ]);
      * ?>
      * ```
-     * 
+     *
      * @param $formSelector
      * @param $params
      * @see \Codeception\Lib\InnerBrowser::dontSeeInFormFields()
@@ -1072,16 +1226,29 @@ trait AcceptanceTesterActions
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * Submits the given form on the page, optionally with the given form
-     * values.  Give the form fields values as an array.
+     * Submits the given form on the page, with the given form
+     * values.  Pass the form field's values as an array in the second
+     * parameter.
      *
-     * Skipped fields will be filled by their values from the page.
+     * Although this function can be used as a short-hand version of
+     * `fillField()`, `selectOption()`, `click()` etc. it has some important
+     * differences:
+     *
+     *  * Only field *names* may be used, not CSS/XPath selectors nor field labels
+     *  * If a field is sent to this function that does *not* exist on the page,
+     *    it will silently be added to the HTTP request.  This is helpful for testing
+     *    some types of forms, but be aware that you will *not* get an exception
+     *    like you would if you called `fillField()` or `selectOption()` with
+     *    a missing field.
+     *
+     * Fields that are not provided will be filled by their values from the page,
+     * or from any previous calls to `fillField()`, `selectOption()` etc.
      * You don't need to click the 'Submit' button afterwards.
      * This command itself triggers the request to form's action.
      *
-     * You can optionally specify what button's value to include
-     * in the request with the last parameter as an alternative to
-     * explicitly setting its value in the second parameter, as
+     * You can optionally specify which button's value to include
+     * in the request with the last parameter (as an alternative to
+     * explicitly setting its value in the second parameter), as
      * button values are not otherwise included in the request.
      *
      * Examples:
@@ -1137,10 +1304,10 @@ trait AcceptanceTesterActions
      * ```
      * Note that "2" will be the submitted value for the "plan" field, as it is
      * the selected option.
-     * 
+     *
      * You can also emulate a JavaScript submission by not specifying any
      * buttons in the third parameter to submitForm.
-     * 
+     *
      * ```php
      * <?php
      * $I->submitForm(
@@ -1154,9 +1321,10 @@ trait AcceptanceTesterActions
      *     ]
      * );
      * ```
-     * 
-     * Pair this with seeInFormFields for quick testing magic.
-     * 
+     *
+     * This function works well when paired with `seeInFormFields()`
+     * for quickly testing CRUD interfaces and form validation logic.
+     *
      * ``` php
      * <?php
      * $form = [
@@ -1165,15 +1333,14 @@ trait AcceptanceTesterActions
      *      'checkbox1' => true,
      *      // ...
      * ];
-     * $I->submitForm('//form[@id=my-form]', $form, 'submitButton');
+     * $I->submitForm('#my-form', $form, 'submitButton');
      * // $I->amOnPage('/path/to/form-page') may be needed
-     * $I->seeInFormFields('//form[@id=my-form]', $form);
-     * ?>
+     * $I->seeInFormFields('#my-form', $form);
      * ```
      *
      * Parameter values can be set to arrays for multiple input fields
      * of the same name, or multi-select combo boxes.  For checkboxes,
-     * either the string value can be used, or boolean values which will
+     * you can use either the string value or boolean `true`/`false` which will
      * be replaced by the checkbox's value in the DOM.
      *
      * ``` php
@@ -1182,7 +1349,7 @@ trait AcceptanceTesterActions
      *      'field1' => 'value',
      *      'checkbox' => [
      *          'value of first checkbox',
-     *          'value of second checkbox,
+     *          'value of second checkbox',
      *      ],
      *      'otherCheckboxes' => [
      *          true,
@@ -1194,27 +1361,29 @@ trait AcceptanceTesterActions
      *          'second option value'
      *      ]
      * ]);
-     * ?>
      * ```
      *
      * Mixing string and boolean values for a checkbox's value is not supported
      * and may produce unexpected results.
-     * 
-     * Field names ending in "[]" must be passed without the trailing square 
+     *
+     * Field names ending in `[]` must be passed without the trailing square
      * bracket characters, and must contain an array for its value.  This allows
      * submitting multiple values with the same name, consider:
-     * 
+     *
      * ```php
+     * <?php
+     * // This will NOT work correctly
      * $I->submitForm('#my-form', [
      *     'field[]' => 'value',
-     *     'field[]' => 'another value', // 'field[]' is already a defined key
+     *     'field[]' => 'another value',  // 'field[]' is already a defined key
      * ]);
      * ```
-     * 
+     *
      * The solution is to pass an array value:
-     * 
+     *
      * ```php
-     * // this way both values are submitted
+     * <?php
+     * // This way both values are submitted
      * $I->submitForm('#my-form', [
      *     'field' => [
      *         'value',
@@ -1222,7 +1391,7 @@ trait AcceptanceTesterActions
      *     ]
      * ]);
      * ```
-     * 
+     *
      * @param $selector
      * @param $params
      * @param $button
@@ -1272,6 +1441,15 @@ trait AcceptanceTesterActions
      * ``` php
      * <?php
      * $I->selectOption('Which OS do you use?', array('Windows','Linux'));
+     * ?>
+     * ```
+     *
+     * Or provide an associative array for the second argument to specifically define which selection method should be used:
+     *
+     * ``` php
+     * <?php
+     * $I->selectOption('Which OS do you use?', array('text' => 'Windows')); // Only search by text 'Windows'
+     * $I->selectOption('Which OS do you use?', array('value' => 'windows')); // Only search by value 'windows'
      * ?>
      * ```
      *
@@ -1325,7 +1503,7 @@ trait AcceptanceTesterActions
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
-     * Attaches a file relative to the Codeception data directory to the given file upload field.
+     * Attaches a file relative to the Codeception `_data` directory to the given file upload field.
      *
      * ``` php
      * <?php
@@ -1418,7 +1596,8 @@ trait AcceptanceTesterActions
      * [!] Method is generated. Documentation taken from corresponding module.
      *
      * Finds and returns the text contents of the given element.
-     * If a fuzzy locator is used, the element is found using CSS, XPath, and by matching the full page source by regular expression.
+     * If a fuzzy locator is used, the element is found using CSS, XPath,
+     * and by matching the full page source by regular expression.
      *
      * ``` php
      * <?php
@@ -1453,7 +1632,7 @@ trait AcceptanceTesterActions
      *
      * @param $cssOrXpath
      * @param $attribute
-     * @internal param $element
+     *
      * @return mixed
      * @see \Codeception\Lib\InnerBrowser::grabAttributeFrom()
      */
@@ -1465,7 +1644,28 @@ trait AcceptanceTesterActions
     /**
      * [!] Method is generated. Documentation taken from corresponding module.
      *
+     * Grabs either the text content, or attribute values, of nodes
+     * matched by $cssOrXpath and returns them as an array.
      *
+     * ```html
+     * <a href="#first">First</a>
+     * <a href="#second">Second</a>
+     * <a href="#third">Third</a>
+     * ```
+     *
+     * ```php
+     * <?php
+     * // would return ['First', 'Second', 'Third']
+     * $aLinkText = $I->grabMultiple('a');
+     *
+     * // would return ['#first', '#second', '#third']
+     * $aLinks = $I->grabMultiple('a', 'href');
+     * ?>
+     * ```
+     *
+     * @param $cssOrXpath
+     * @param $attribute
+     * @return string[]
      * @see \Codeception\Lib\InnerBrowser::grabMultiple()
      */
     public function grabMultiple($cssOrXpath, $attribute = null) {
@@ -1490,7 +1690,7 @@ trait AcceptanceTesterActions
      * [!] Method is generated. Documentation taken from corresponding module.
      *
      * Sets a cookie with the given name and value.
-     * You can set additional cookie params like `domain`, `path`, `expire`, `secure` in array passed as last argument.
+     * You can set additional cookie params like `domain`, `path`, `expires`, `secure` in array passed as last argument.
      *
      * ``` php
      * <?php
@@ -1524,6 +1724,21 @@ trait AcceptanceTesterActions
      */
     public function grabCookie($cookie, $params = null) {
         return $this->getScenario()->runStep(new \Codeception\Step\Action('grabCookie', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Grabs current page source code.
+     *
+     * @throws ModuleException if no page was opened.
+     *
+     * @return string Current page source code.
+     * @see \Codeception\Lib\InnerBrowser::grabPageSource()
+     */
+    public function grabPageSource() {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('grabPageSource', func_get_args()));
     }
 
  
@@ -1730,13 +1945,11 @@ trait AcceptanceTesterActions
      * ``` php
      * <?php
      * $I->seeNumberOfElements('tr', 10);
-     * $I->seeNumberOfElements('tr', [0,10]); //between 0 and 10 elements
+     * $I->seeNumberOfElements('tr', [0,10]); // between 0 and 10 elements
      * ?>
      * ```
      * @param $selector
-     * @param mixed $expected :
-     * - string: strict number
-     * - array: range of numbers [0,10]
+     * @param mixed $expected int or int[]
      * Conditional Assertion: Test won't be stopped on fail
      * @see \Codeception\Lib\InnerBrowser::seeNumberOfElements()
      */
@@ -1751,13 +1964,11 @@ trait AcceptanceTesterActions
      * ``` php
      * <?php
      * $I->seeNumberOfElements('tr', 10);
-     * $I->seeNumberOfElements('tr', [0,10]); //between 0 and 10 elements
+     * $I->seeNumberOfElements('tr', [0,10]); // between 0 and 10 elements
      * ?>
      * ```
      * @param $selector
-     * @param mixed $expected :
-     * - string: strict number
-     * - array: range of numbers [0,10]
+     * @param mixed $expected int or int[]
      * @see \Codeception\Lib\InnerBrowser::seeNumberOfElements()
      */
     public function seeNumberOfElements($selector, $expected) {
@@ -1877,9 +2088,15 @@ trait AcceptanceTesterActions
      *
      * Checks that response code is equal to value provided.
      *
-     * @param $code
+     * ```php
+     * <?php
+     * $I->seeResponseCodeIs(200);
      *
-     * @return mixed
+     * // recommended \Codeception\Util\HttpCode
+     * $I->seeResponseCodeIs(\Codeception\Util\HttpCode::OK);
+     * ```
+     *
+     * @param $code
      * Conditional Assertion: Test won't be stopped on fail
      * @see \Codeception\Lib\InnerBrowser::seeResponseCodeIs()
      */
@@ -1891,13 +2108,169 @@ trait AcceptanceTesterActions
      *
      * Checks that response code is equal to value provided.
      *
-     * @param $code
+     * ```php
+     * <?php
+     * $I->seeResponseCodeIs(200);
      *
-     * @return mixed
+     * // recommended \Codeception\Util\HttpCode
+     * $I->seeResponseCodeIs(\Codeception\Util\HttpCode::OK);
+     * ```
+     *
+     * @param $code
      * @see \Codeception\Lib\InnerBrowser::seeResponseCodeIs()
      */
     public function seeResponseCodeIs($code) {
         return $this->getScenario()->runStep(new \Codeception\Step\Assertion('seeResponseCodeIs', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that response code is between a certain range. Between actually means [from <= CODE <= to]
+     *
+     * @param $from
+     * @param $to
+     * Conditional Assertion: Test won't be stopped on fail
+     * @see \Codeception\Lib\InnerBrowser::seeResponseCodeIsBetween()
+     */
+    public function canSeeResponseCodeIsBetween($from, $to) {
+        return $this->getScenario()->runStep(new \Codeception\Step\ConditionalAssertion('seeResponseCodeIsBetween', func_get_args()));
+    }
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that response code is between a certain range. Between actually means [from <= CODE <= to]
+     *
+     * @param $from
+     * @param $to
+     * @see \Codeception\Lib\InnerBrowser::seeResponseCodeIsBetween()
+     */
+    public function seeResponseCodeIsBetween($from, $to) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Assertion('seeResponseCodeIsBetween', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that response code is equal to value provided.
+     *
+     * ```php
+     * <?php
+     * $I->dontSeeResponseCodeIs(200);
+     *
+     * // recommended \Codeception\Util\HttpCode
+     * $I->dontSeeResponseCodeIs(\Codeception\Util\HttpCode::OK);
+     * ```
+     * @param $code
+     * Conditional Assertion: Test won't be stopped on fail
+     * @see \Codeception\Lib\InnerBrowser::dontSeeResponseCodeIs()
+     */
+    public function cantSeeResponseCodeIs($code) {
+        return $this->getScenario()->runStep(new \Codeception\Step\ConditionalAssertion('dontSeeResponseCodeIs', func_get_args()));
+    }
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that response code is equal to value provided.
+     *
+     * ```php
+     * <?php
+     * $I->dontSeeResponseCodeIs(200);
+     *
+     * // recommended \Codeception\Util\HttpCode
+     * $I->dontSeeResponseCodeIs(\Codeception\Util\HttpCode::OK);
+     * ```
+     * @param $code
+     * @see \Codeception\Lib\InnerBrowser::dontSeeResponseCodeIs()
+     */
+    public function dontSeeResponseCodeIs($code) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Assertion('dontSeeResponseCodeIs', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that the response code 2xx
+     * Conditional Assertion: Test won't be stopped on fail
+     * @see \Codeception\Lib\InnerBrowser::seeResponseCodeIsSuccessful()
+     */
+    public function canSeeResponseCodeIsSuccessful() {
+        return $this->getScenario()->runStep(new \Codeception\Step\ConditionalAssertion('seeResponseCodeIsSuccessful', func_get_args()));
+    }
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that the response code 2xx
+     * @see \Codeception\Lib\InnerBrowser::seeResponseCodeIsSuccessful()
+     */
+    public function seeResponseCodeIsSuccessful() {
+        return $this->getScenario()->runStep(new \Codeception\Step\Assertion('seeResponseCodeIsSuccessful', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that the response code 3xx
+     * Conditional Assertion: Test won't be stopped on fail
+     * @see \Codeception\Lib\InnerBrowser::seeResponseCodeIsRedirection()
+     */
+    public function canSeeResponseCodeIsRedirection() {
+        return $this->getScenario()->runStep(new \Codeception\Step\ConditionalAssertion('seeResponseCodeIsRedirection', func_get_args()));
+    }
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that the response code 3xx
+     * @see \Codeception\Lib\InnerBrowser::seeResponseCodeIsRedirection()
+     */
+    public function seeResponseCodeIsRedirection() {
+        return $this->getScenario()->runStep(new \Codeception\Step\Assertion('seeResponseCodeIsRedirection', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that the response code is 4xx
+     * Conditional Assertion: Test won't be stopped on fail
+     * @see \Codeception\Lib\InnerBrowser::seeResponseCodeIsClientError()
+     */
+    public function canSeeResponseCodeIsClientError() {
+        return $this->getScenario()->runStep(new \Codeception\Step\ConditionalAssertion('seeResponseCodeIsClientError', func_get_args()));
+    }
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that the response code is 4xx
+     * @see \Codeception\Lib\InnerBrowser::seeResponseCodeIsClientError()
+     */
+    public function seeResponseCodeIsClientError() {
+        return $this->getScenario()->runStep(new \Codeception\Step\Assertion('seeResponseCodeIsClientError', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that the response code is 5xx
+     * Conditional Assertion: Test won't be stopped on fail
+     * @see \Codeception\Lib\InnerBrowser::seeResponseCodeIsServerError()
+     */
+    public function canSeeResponseCodeIsServerError() {
+        return $this->getScenario()->runStep(new \Codeception\Step\ConditionalAssertion('seeResponseCodeIsServerError', func_get_args()));
+    }
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Checks that the response code is 5xx
+     * @see \Codeception\Lib\InnerBrowser::seeResponseCodeIsServerError()
+     */
+    public function seeResponseCodeIsServerError() {
+        return $this->getScenario()->runStep(new \Codeception\Step\Assertion('seeResponseCodeIsServerError', func_get_args()));
     }
 
  
@@ -1968,5 +2341,42 @@ trait AcceptanceTesterActions
      */
     public function dontSeeInTitle($title) {
         return $this->getScenario()->runStep(new \Codeception\Step\Assertion('dontSeeInTitle', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Switch to iframe or frame on the page.
+     *
+     * Example:
+     * ``` html
+     * <iframe name="another_frame" src="http://example.com">
+     * ```
+     *
+     * ``` php
+     * <?php
+     * # switch to iframe
+     * $I->switchToIframe("another_frame");
+     * ```
+     *
+     * @param string $name
+     * @see \Codeception\Lib\InnerBrowser::switchToIframe()
+     */
+    public function switchToIframe($name) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('switchToIframe', func_get_args()));
+    }
+
+ 
+    /**
+     * [!] Method is generated. Documentation taken from corresponding module.
+     *
+     * Moves back in history.
+     *
+     * @param int $numberOfSteps (default value 1)
+     * @see \Codeception\Lib\InnerBrowser::moveBack()
+     */
+    public function moveBack($numberOfSteps = null) {
+        return $this->getScenario()->runStep(new \Codeception\Step\Action('moveBack', func_get_args()));
     }
 }
