@@ -1,11 +1,9 @@
-<?php  //[STAMP] ffada6587f597b23930f2f63d3309fd0
+<?php  //[STAMP] aa740ba49ecc9060e057de5a4441b25b
 namespace _generated;
 
 // This class was automatically generated by build task
 // You should not change it manually as it will be overwritten on next build
 // @codingStandardsIgnoreFile
-
-use Helper\Functional;
 
 trait FunctionalTesterActions
 {
